Added bomb support
Bomb support sucks - currently adds a BombEntity to the map, which
detonates, and adds FireEntities to the map - we need to re-work how we
keep track of entities on both the client and server side (probably via
UUIDs of their own).  We also need to re-think how we approach redraws -
maybe a game state should tell the graphics render system what needs to
be redrawn?